refactor(o_minimal/definable): don't bundle definable types
Bundling made it far too difficult to express definability of
operations like addition. It also avoids the need for `X.prod Y`.
By not bundling we follow the pattern used by mathlib
(in topology for example).

A side effect is that now the structure `S` must be
an explicit argument to `def_set` and `def_fun`.